# [automation] - Consume config set-all and rm-all commands