### Go Benchmark Results

**Before**

- Mixed (10 secret, 10 plain): 11.69s
- Mixed (20 secret, 20 plain): 24.85s
- Plain (20 plain): 11.25s
- Plain (40 plain): 26.39s
- Secret (20 secret): 12.29s
- Secret (40 secret): 25.36s

**After**

- Mixed (10 secret, 10 plain): 1.18s
- Mixed (20 secret, 20 plain): 1.55s
- Plain (20 plain): 1.05s
- Plain (40 plain): 1.20s
- Secret (20 secret): 1.51s
- Secret (40 secret): 2.54s

Related to: https://github.com/pulumi/pulumi/pull/6373